+++ silence/server/endpoint_loader.py
@@ -5,6 +5,7 @@ from os.path import splitext
 
 from silence.server import manager, default_endpoints
 from silence.settings import settings
+from silence.logging.default_logger import logger
 
 ###############################################################################
 # Look for .py files inside the project's "/api" folder
@@ -12,9 +13,12 @@ from silence.settings import settings
 ###############################################################################
 
 def load_user_endpoints():
+    logger.debug("Looking for custom endpoints...")
+
     # Load every .py file inside the api/ folder
     for pyfile in listdir("api"):
         module_name = "api." + splitext(pyfile)[0]
+        logger.debug(f"Found endpoint file: {module_name}")
 
         try:
             importlib.import_module(module_name)
@@ -29,5 +33,8 @@ def load_default_endpoints():
     if route_prefix.endswith("/"):
         route_prefix = route_prefix[:-1]
 
-    manager.APP.add_url_rule(f"{route_prefix}/login", "login", default_endpoints.login, methods=["POST"])
-    manager.APP.add_url_rule(f"{route_prefix}/register", "register", default_endpoints.register, methods=["POST"])
+    if settings.ENABLE_LOGIN:
+        manager.APP.add_url_rule(f"{route_prefix}/login", "login", default_endpoints.login, methods=["POST"])
+
+    if settings.ENABLE_REGISTER:
+        manager.APP.add_url_rule(f"{route_prefix}/register", "register", default_endpoints.register, methods=["POST"])