+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.4.0
+- Added support for application/json as encoding for POST/PUT requests
+- Added configuration options to disable login and register
+- Custom formatting for HTTP log messages
+- Use a connection pool to connect to the DB
+
 # 0.3.0
 - Added support for the default endpoints /login and /register using custom tables and fields
 - Added support for session tokens and authentication requirements